+++ history.md
@@ -1,5 +1,6 @@
 [2021/12/28]
 - Add
+  + open.commons.utils.ArrayUtils.objectArray(Object, Object[])
   + open.commons.utils.ArrayUtils.objectArray(Object[], Object...)
   
 [2021/12/20]